feat: impl import config, refactor
