+++ Cargo.toml
@@ -15,6 +15,7 @@ reqwest = { version = "0.12.2", features = [
     "blocking",
     "cookies",
 ] }
+base64 = "0.22.0"
 clap = { version = "4.5.4", features = ["derive"] }
 tokio = { version = "1.37.0", features = ["full"] }
 tokio-serde = { version = "0.9.0", features = ["json", "serde"] }
